Create PhaserGame component
