+++ src/phaser/index.tsx
@@ -1,23 +1,3 @@
-import { AUTO, Game } from "phaser"
-import { Boot } from "./scenes/Boot"
-import { GameOver } from "./scenes/GameOver"
-import { Game as MainGame } from "./scenes/Game"
-import { MainMenu } from "./scenes/MainMenu"
-import { Preloader } from "./scenes/Preloader"
-
-//  Find out more information about the Game Config at:
-//  https://docs.phaser.io/api-documentation/typedef/types-core#gameconfig
-const config: Phaser.Types.Core.GameConfig = {
-  type: AUTO,
-  width: 1024,
-  height: 768,
-  parent: "game-container",
-  backgroundColor: "#028af8",
-  scene: [Boot, Preloader, MainMenu, MainGame, GameOver],
-}
-
-const StartGame = (parent: string) => {
-  return new Game({ ...config, parent })
-}
-
-export default StartGame
+// WARN: Do not import Phaser or any file that statically imports Phaser as it
+// will be imported by our SSR code and break the build!
+export { default as PhaserGame, type PhaserGameProps } from "./PhaserGame"
