+++ src/phaser/scenes/index.ts
@@ -1,0 +1,5 @@
+export { Boot } from "./Boot"
+export { Preloader } from "./Preloader"
+export { MainMenu } from "./MainMenu"
+export { Game } from "./Game"
+export { GameOver } from "./GameOver"
